lfg-loan-backend/requirements.txt
Django
djangorestframework
drf-yasg
django-cors-headers
psycopg2-binary
djangorestframework_simplejwt
daphne
redis
celery[redis]
channels
channels-redis
asgiref
wheel
django-filter
python-dotenv
python-magic
requests
gunicorn